refactor(): separate enums and interfaces to according files
