+++ db.json
@@ -1,4 +1,164 @@
 {
   "graphs": [
+    {
+      "cells": [
+        {
+          "type": "app.NodeShape",
+          "position": {
+            "x": 1190,
+            "y": 1310
+          },
+          "size": {
+            "width": 45,
+            "height": 45
+          },
+          "angle": 0,
+          "id": "474ca274-5c6e-4834-8bf9-3b560f27ff6e",
+          "localType": "NodeShape",
+          "ports": {
+            "groups": {
+              "in": {
+                "position": {
+                  "name": "left"
+                },
+                "attrs": {
+                  "portBody": {
+                    "magnet": "passive",
+                    "r": 4,
+                    "fill": "darkBlue"
+                  }
+                },
+                "markup": [
+                  {
+                    "tagName": "circle",
+                    "selector": "portBody"
+                  }
+                ]
+              },
+              "out": {
+                "position": {
+                  "name": "right"
+                },
+                "attrs": {
+                  "portBody": {
+                    "magnet": true,
+                    "r": 4,
+                    "fill": "darkRed"
+                  }
+                },
+                "markup": [
+                  {
+                    "tagName": "circle",
+                    "selector": "portBody"
+                  }
+                ]
+              }
+            },
+            "items": [
+              {
+                "group": "in",
+                "id": "e8dbf25f-5405-423f-a55a-13a524ccf564"
+              },
+              {
+                "group": "out",
+                "id": "a0ca27ed-5cc2-4da0-850e-84fe39b884de"
+              }
+            ]
+          },
+          "z": 1,
+          "attrs": {
+            "label": {
+              "text": "Kinesis Stream",
+              "textAnchor": "middle",
+              "refX": "50%",
+              "refY": "110%",
+              "fontSize": 10
+            },
+            "icon": {
+              "href": "icons/Arch_Analytics/Arch_Amazon-Kinesis_64.svg"
+            }
+          }
+        },
+        {
+          "type": "app.NodeShape",
+          "position": {
+            "x": 1380,
+            "y": 1200
+          },
+          "size": {
+            "width": 45,
+            "height": 45
+          },
+          "angle": 0,
+          "id": "9055eed2-3339-4f0a-83d5-1b3513e0a948",
+          "localType": "NodeShape",
+          "ports": {
+            "groups": {
+              "in": {
+                "position": {
+                  "name": "left"
+                },
+                "attrs": {
+                  "portBody": {
+                    "magnet": "passive",
+                    "r": 4,
+                    "fill": "darkBlue"
+                  }
+                },
+                "markup": [
+                  {
+                    "tagName": "circle",
+                    "selector": "portBody"
+                  }
+                ]
+              },
+              "out": {
+                "position": {
+                  "name": "right"
+                },
+                "attrs": {
+                  "portBody": {
+                    "magnet": true,
+                    "r": 4,
+                    "fill": "darkRed"
+                  }
+                },
+                "markup": [
+                  {
+                    "tagName": "circle",
+                    "selector": "portBody"
+                  }
+                ]
+              }
+            },
+            "items": [
+              {
+                "group": "in",
+                "id": "e8dbf25f-5405-423f-a55a-13a524ccf564"
+              },
+              {
+                "group": "out",
+                "id": "a0ca27ed-5cc2-4da0-850e-84fe39b884de"
+              }
+            ]
+          },
+          "z": 2,
+          "attrs": {
+            "label": {
+              "text": "Redshift",
+              "textAnchor": "middle",
+              "refX": "50%",
+              "refY": "110%",
+              "fontSize": 10
+            },
+            "icon": {
+              "href": "icons/Arch_Analytics/Arch_Amazon-Redshift_64.svg"
+            }
+          }
+        }
+      ],
+      "diagramName": "Diagram Name",
+      "id": 1
+    }
   ]
 }
