+++ src/app/components/NavLink/NavLink.tsx
@@ -1,10 +1,7 @@
-import { Link, LinkProps } from "react-router-dom";
+import { Link } from "react-router-dom";
 import styles from "./navLink.module.scss";
 import React from "react";
-
-interface NavLinkProps extends LinkProps {
-    customClassName?: string;
-}
+import { NavLinkProps } from "../../types/interfaces";
 
 const NavLink = ({customClassName, ...linkProps}: NavLinkProps) => {
     return <Link className={`${styles.link} ${customClassName}`} {...linkProps} />;
