+++ src/app/types/interfaces.ts
@@ -1,3 +1,5 @@
+import { LinkProps } from "react-router-dom";
+
 import { dia } from "@clientio/rappid";
 
 export {}
@@ -18,4 +20,13 @@ export interface MaxLinksInputProps {
 
 export interface ColorInputProps {
     cell: dia.Cell
+}
+
+export interface NavLinkProps extends LinkProps {
+    customClassName?: string;
+}
+
+export interface JSONGraphRootState {
+    diagramList: [{ cells: [], id: number, diagramName: string }] | null
+    id: number | null
 }
